+++ packages/rev-ui/src/views/__tests__/ListView.tests.tsx
@@ -423,7 +423,7 @@ describe('ListView', () => {
     describe('component props - where clause', () => {
         const fieldList = ['id', 'title', 'published', 'post_date'];
         let meta: IModelMeta<models.Post>;
-        const whereClause = { id: { _gt: 5 }};
+        const whereClause = { id: { _gte: 5 }};
         let expectedData: models.Post[];
         let expectedFields: fields.Field[];
 
@@ -465,6 +465,57 @@ describe('ListView', () => {
         });
 
         it('passes the correct record data', () => {
+            expectedData.forEach((record, recordIdx) => {
+                fieldList.forEach((fieldName) => {
+                    expect(receivedProps.records[recordIdx][fieldName])
+                        .to.equal(record[fieldName]);
+                });
+            });
+        });
+
+    });
+
+    describe('component props - orderBy', () => {
+        const fieldList = ['id', 'title', 'published', 'post_date'];
+        let meta: IModelMeta<models.Post>;
+        const orderBy = ['title desc', 'published'];
+        let expectedData: models.Post[];
+        let expectedFields: fields.Field[];
+
+        before(async () => {
+            modelManager = models.getModelManager();
+            meta = modelManager.getModelMeta(models.Post);
+            await createData(modelManager);
+            expectedData = (await modelManager.read(models.Post, {
+                orderBy: orderBy
+            })).results;
+            expectedFields = [
+                meta.fieldsByName['id'],
+                meta.fieldsByName['title'],
+                meta.fieldsByName['published'],
+                meta.fieldsByName['post_date']
+            ];
+
+            lifecycleOptions.enableComponentDidMount = true;
+            receivedProps = null;
+            const wrapper = mountComponent(
+                <ListView
+                    title="List with Data Loaded..."
+                    model={model}
+                    fields={fieldList}
+                    orderBy={orderBy}
+                    limit={100}
+                    component={SpyComponent} />
+            );
+            await sleep(10);
+            wrapper.update();
+        });
+
+        it('passes the fields list', () => {
+            expect(receivedProps.fields).to.deep.equal(expectedFields);
+        });
+
+        it('returns records in the expected order', () => {
             expectedData.forEach((record, recordIdx) => {
                 fieldList.forEach((fieldName) => {
                     expect(receivedProps.records[recordIdx][fieldName])